requirements.txt
plotly==5.3.1
pillow==8.4.0
pandas==1.3.4
pydantic==1.8.2
kaleido==0.2.1
grpcio==1.41.1
protobuf==3.19.1